movimiento habilitado, falta aproximar bien la llegada de los moviles a las intersecciones

+++ src/objetosTransito/Comparendo.scala
@@ -2,7 +2,7 @@ package objetosTransito
 import movil.Vehiculo
 import scala.collection.mutable.ArrayBuffer
 import rectas.Via
-protected class Comparendo(vehiculo : Vehiculo,velVehiculo : Int,velMax : Int, nombreVia : String) {
+protected class Comparendo(vehiculo : Vehiculo,velVehiculo :Double,velMax :Double, nombreVia : String) {
   
 }
